[HTML5] Fix for memory leak when not using webgl and using sprite_delete
https://github.com/YoYoGames/GameMaker-Bugs/issues/7206

+++ scripts/yySprite.js
@@ -1,4 +1,4 @@
-// **********************************************************************************************************************
+﻿// **********************************************************************************************************************
 // 
 // Copyright (c)2011, YoYo Games Ltd. All Rights reserved.
 // 
@@ -2781,17 +2781,28 @@ yySpriteManager.prototype.Delete = function(_id) {
 			var pTPE = pSprite.ppTPE[i];
 			if (!pTPE || !pTPE.texture) continue;
 			var pTexture = pTPE.texture;
-			if (!pTexture || !pTexture.webgl_textureid) continue;
-			if (flush) {
-				g_webGL.FlushAll();
-				flush = false;
-			}
-			
-			// Update state manager here 
-			g_webGL.RSMan.ClearTexture(pTexture.webgl_textureid); 
 
-			g_webGL.DeleteTexture(pTexture.webgl_textureid.Texture);
-			pTexture.webgl_textureid = null;
+			if ((g_pGMFile.Options.WebGL) && (g_pGMFile.Options.WebGL != 0))
+			{
+
+				if (!pTexture || !pTexture.webgl_textureid) continue;
+				if (flush) {
+					g_webGL.FlushAll();
+					flush = false;
+				}
+				
+				// Update state manager here 
+				g_webGL.RSMan.ClearTexture(pTexture.webgl_textureid); 
+
+				g_webGL.DeleteTexture(pTexture.webgl_textureid.Texture);
+				pTexture.webgl_textureid = null;
+			}
+			else
+			{
+				g_Textures[pTPE.texture]=null;
+				pTPE.texture = null;
+					 
+			}
 		}
 		this.Sprites[_id] = undefined;
 	}
